refactor: community views 파일 분리
